+++ ChangeLog.md
@@ -2,6 +2,7 @@
 + CI
   - Add code spell
   - Add code QL analysis
+  - Add details in CDlgAbout
 
 ### Version: v1.0.9
 + Add CBrowseFolder and CTitleBar